backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-admin-logs==1.3.0
django-environ==0.11.2
django-unfold==0.42.0
gunicorn==23.0.0
packaging==24.2
sqlparse==0.5.2
tzdata==2024.2
whitenoise==6.8.2
